Minor release [v4.4.0.0]
- Added a settings menu
- Added a QoL feature that skips the download process automatically if the raw file is detected (this is helpful if the user chose an undesired format and wants to change it without redownloading the entire video)
- Added a migration feature to prevent the loss of user config data when loading from older versions
- Greatly optimized the queue and history storage systems, it now stores each in a single compressed file
- Small UI improvements

+++ src/MediaDownloader/MediaDownloader/Scripts/Tools/Files.cs
@@ -4,12 +4,38 @@ namespace MediaDownloader.Tools
 {
     public static class Files
     {
-        public static void CleanFolder(string path)
+        public static void CleanPath(string path, string[] ignoredExtensions)
+        {
+            CleanFiles(path, ignoredExtensions);
+            CleanFolders(path);
+        }
+
+        public static void CleanFiles(string path, string[] ignoredFiles)
         {
             // delete all files
             foreach (string file in Directory.GetFiles(path))
-                File.Delete(file);
+            {
+                bool deleteFile = true;
+                if (ignoredFiles != null)
+                {
+                    for (int i = 0; i < ignoredFiles.Length; i++)
+                    {
+                        string fileName = Path.GetFileName(file);
+                        if (fileName == ignoredFiles[i])
+                        {
+                            deleteFile = false;
+                            break;
+                        }
+                    }
+                }
 
+                if (deleteFile)
+                    File.Delete(file);
+            }
+        }
+
+        public static void CleanFolders(string path)
+        {
             // delete all folders
             foreach (string folder in Directory.GetDirectories(path))
                 Directory.Delete(folder, true);
